+++ lib/Container/ContainerBuilder.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Phpactor\Extension;
+namespace Phpactor\Container;
 
 use Closure;
+use Phpactor\Container\Container;
 
 interface ContainerBuilder
 {
