+++ lib/Container/Extension.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Phpactor\Extension;
+namespace Phpactor\Container;
 
 use PhpBench\DependencyInjection\ExtensionInterface;
+use Phpactor\Container\ContainerBuilder;
 
 interface Extension
 {
