+++ lib/Container/Schema.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace Phpactor\Extension;
+namespace Phpactor\Container;
+
+use Phpactor\Container\InvalidConfig;
+use Phpactor\Container\Schema;
 
 class Schema
 {
