+++ lib/Extension/ClassMover/ClassMoverExtension.php
@@ -12,10 +12,10 @@ use Phpactor\ClassMover\Adapter\WorseTolerant\WorseTolerantMemberReplacer;
 use Phpactor\ClassMover\ClassMover;
 use Phpactor\Extension\ClassMover\Command\ClassCopyCommand;
 use Phpactor\Extension\ClassMover\Command\ClassMoveCommand;
-use Phpactor\Extension\ContainerBuilder;
-use Phpactor\Extension\Extension;
-use Phpactor\Extension\Schema;
-use Phpactor\Extension\Container;
+use Phpactor\Container\ContainerBuilder;
+use Phpactor\Container\Extension;
+use Phpactor\Container\Schema;
+use Phpactor\Container\Container;
 use Phpactor\Extension\ClassMover\Command\ReferencesMemberCommand;
 use Phpactor\Extension\ClassMover\Command\ReferencesClassCommand;
 use Phpactor\Extension\ClassMover\Application\ClassMemberReferences;
