+++ lib/Extension/ClassToFile/ClassToFileExtension.php
@@ -2,9 +2,9 @@
 
 namespace Phpactor\Extension\ClassToFile;
 
-use Phpactor\Extension\Extension;
-use Phpactor\Extension\ContainerBuilder;
-use Phpactor\Extension\Schema;
+use Phpactor\Container\Extension;
+use Phpactor\Container\ContainerBuilder;
+use Phpactor\Container\Schema;
 
 class ClassToFileExtension implements Extension
 {
