+++ lib/Extension/PathFinder/PathFinderExtension.php
@@ -3,10 +3,10 @@
 namespace Phpactor\Extension\PathFinder;
 
 use Phpactor\ClassFileConverter\PathFinder;
-use Phpactor\Extension\Extension;
-use Phpactor\Extension\ContainerBuilder;
-use Phpactor\Extension\Schema;
-use Phpactor\Extension\Container;
+use Phpactor\Container\Extension;
+use Phpactor\Container\ContainerBuilder;
+use Phpactor\Container\Schema;
+use Phpactor\Container\Container;
 use Phpactor\Extension\PathFinder\Application\Navigator;
 
 class PathFinderExtension implements Extension
